+++ src/views/legal/MentionsLegalesView.vue
@@ -40,7 +40,7 @@
 
         <div
           data-testid="mentions-legales-content"
-          class="bg-white dark:bg-gray-800 rounded-lg shadow-md p-6 tablet:p-10 space-y-8"
+          class="bg-white dark:bg-gray-800 text-slate-800 dark:text-white rounded-lg shadow-md p-6 tablet:p-10 space-y-8"
         >
           <p>{{ t('Legal.Mentions.Preamble') }}</p>